real-estate-newsletter/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.18.1
python-dotenv==1.0.0
lxml==4.9.3
urllib3==2.0.4
selenium==4.15.0
